(longlong) Implement shifts.
Promote to using int64_t internally since we need to (still some
missing)

+++ src/sccz80/declparse.c
@@ -202,7 +202,7 @@ void free_type(void *data)
     FREENULL(type);
 }
 
-Type *make_constant(const char *name, int32_t value)
+Type *make_constant(const char *name, int64_t value)
 {
     SYMBOL *ptr;
     Type *type = CALLOC(1,sizeof(*type));
